+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Added `I_TopologyJointRule` in GH components to support TOPO_I joints.
+
 ### Changed
 
 ### Removed